fix csv output
